Déplacement de la déconnexion dans Welcome.js ++ Redirection HeaderKid + HeaderParent vers Welcomes.js
