+++ resources/js/views/Welcome/Welcome.js
@@ -1,11 +1,34 @@
-import React, { useState } from 'react';
-import {Link} from 'react-router-dom';
+import React, {useState, useContext} from 'react';
+import axios from 'axios';
+import { AuthDataContext } from '../../components/AuthDataProvider';
+import {Redirect} from 'react-router-dom';
+
 
 const Welcome = props => {
     let familyName = "Kidzou";
 
+    const [redirect, setRedirect] = useState(false);
+    const { onLogout } = useContext(AuthDataContext);
+
+    function logout() {
+        axios.post("/api/auth/logout",{})
+        .then(json => {
+            let authData = {};
+            sessionStorage.setItem('authData', JSON.stringify(authData));
+            console.log(json.data.message);
+
+        }).then(onLogout)
+        .catch(error => {
+            console.log(error);
+        });
+
+        setRedirect(true);
+    }
+
+
     return (
         <div className="container background page-welcome">
+            <div className="container__logout" onClick={logout}><i className="fas fa-power-off"></i></div>
             <div className="container__body none">
                 <div className="content  flex-grow">
                     <img className="page-welcome__img" src="/img/welcome-img.svg" alt="Welcome"/>
