+++ resources/sass/base/_container.scss
@@ -71,7 +71,24 @@
       box-shadow: 0 .1rem .1rem $black-shadow;
       color: $white;
       font-weight: bold;
-  }
+    }
+    &__logout {
+      position: absolute;
+      width: fit-content;
+      top: 1rem;
+      right: 1rem;
+      display: flex;
+      text-align: right;
+      font-size: 1.5rem;
+      color: $black;
+      z-index: 200;
+      /*
+      position: relative;
+      width: 100%;
+      display: flex;
+      right: -2.5rem;
+      */
+    }
   }
   
 .content {
