+++ resources/sass/components/_header-profile.scss
@@ -5,6 +5,7 @@
 .page-header-profile {
     width: 100%;
     padding: 0 2.5rem;
+    padding-top: 1rem;
 
     &__logout {
         position: relative;
